libdsound: fix shift-width UB in the DRF field-mask macro
MCPX_REF_MASK computed the field mask as `(1 << width) - 1`. For a
full-width field (width 32, e.g. a `31:0` DRF) that is `1 << 32`, which is
undefined behavior -- the shift count equals the width of int. MSVC
constant-folds it to 0, so the mask came out `0 - 1 == 0xFFFFFFFF`
(correct), which is why retail always worked. clang folds the same UB
differently and produces a garbage mask.

The field that actually bit us is NV_PAPU_FEDECPARAM_VALUE (31:0), the
parameter of a trapped front-end method. On the software-overridden
SE2FE_IDLE_VOICE trap, HandleFETrap decodes the idle hardware-voice index
out of FEDECPARAM via MCPX_GET_REG_VALUE(.., NV_PAPU_FEDECPARAM_VALUE).
With the miscompiled mask the real index (e.g. 0x45) decoded to garbage,
so HandleIdleVoice was handed the wrong voice -- consistently voice 1,
which no client owns. The genuine idling voice therefore never got
RemoveIdleVoice, so every voice sat in the processing list until the
5000ms MCPX_DEBUG_STUCK_VOICES timeout fired RemoveStuckVoice. Under an
app that spams short sounds over a looping music stream this serialised
every teardown behind a 5s stall: audible symptoms were the music
dropping out, short sounds stalling, the music never looping, and the UI
freezing under the storm of 5s waits.

Do the shift in 64-bit: `(DWORD)((1ULL << width) - 1ULL)`. `1ULL << 32`
is well-defined (0x1_0000_0000), giving mask 0xFFFFFFFF; for every width
< 32 the result is bit-identical to before, so no other field changes
behaviour. This also corrects every other 31:0 DRF field decoded through
these macros (memory addresses, SGE context PAs, FIFO values, ...), which
were silently wrong for the same reason.

Verified on hardware (PrometheOS): idle traps now decode the correct
voice index, RemoveIdleVoice runs on the natural path, RemoveStuckVoice
count stays 0, and stream Release drops from ~5000ms to <20ms.

+++ libs/libdsound/dsound/mcpxcore.h
@@ -46,8 +46,15 @@ DEFINETYPE(MCPX_VOICE_HANDLE, WORD);
 #define MCPX_REF_SHIFT(drf) \
     (0 ? drf)
 
+// RXDK: shift in 64-bit. For a full-width field (e.g. 31:0) the width is 32, and
+// the original `1 << 32` is shift-width UB. MSVC constant-folded it to 0 (mask
+// 0xFFFFFFFF, correct); clang folds the UB differently and produces a garbage mask,
+// which mis-decoded NV_PAPU_FEDECPARAM_VALUE (31:0) -- the SE2FE_IDLE_VOICE voice
+// index -- so every idle trap was misrouted and every voice hung on the 5000ms
+// stuck-voice timeout. 1ULL << 32 is well-defined; result is identical for all
+// widths < 32.
 #define MCPX_REF_MASK(drf) \
-    ((1 << ((1 ? drf) - (0 ? drf) + 1)) - 1)
+    ((DWORD)((1ULL << ((1 ? drf) - (0 ? drf) + 1)) - 1ULL))
 
 #define MCPX_CLEAR_REG_VALUE(arg, drf) \
     ((arg) & ~(MCPX_REF_MASK(drf) << MCPX_REF_SHIFT(drf)))
